@@ -60,11 +60,11 @@
     "rollup-watch": "^3.2.2",
     "selenium-server": "^2.53.1",
     "todomvc-app-css": "^2.1.0",
-    "typescript": "^2.1.5",
+    "typescript": "^2.4.2",
     "uglify-js": "^2.7.5",
-    "vue": "^2.3.3",
+    "vue": "DanielRosenwasser/vue#e4a8545",
     "vue-loader": "^12.0.0",
-    "vue-template-compiler": "^2.3.3",
+    "vue-template-compiler": "^2.4.2",
     "webpack": "^2.6.1",
     "webpack-dev-middleware": "^1.10.0",
     "webpack-hot-middleware": "^2.16.1"
